+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotInvalidKeywordOnlyMarkerAnnotation.kt
@@ -7,6 +7,7 @@ import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
 import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle
+import dev.xeonkryptos.xeonrobotframeworkplugin.ide.psi.visitor.EmptyVariableDetector
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLocalArgumentsSetting
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariable
 
@@ -18,10 +19,14 @@ class RobotInvalidKeywordOnlyMarkerAnnotation : Annotator {
         }
         val settingsParent = PsiTreeUtil.getParentOfType(element, RobotLocalArgumentsSetting::class.java)
         if (settingsParent == null) {
-            holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.variable.blank.invalid"))
-                .highlightType(ProblemHighlightType.GENERIC_ERROR)
-                .range(element)
-                .create()
+            val detector = EmptyVariableDetector()
+            element.accept(detector)
+            if (detector.emptyVariable) {
+                holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.variable.blank.invalid"))
+                    .highlightType(ProblemHighlightType.GENERIC_ERROR)
+                    .range(element)
+                    .create()
+            }
         }
     }
 }