+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - Fixed scoping issue with variables leading to incorrectly marked variables as reassigned
 - Multiline handling extended to allow new lines between ellipses
 - Detection/Parsing of IN keywords of FOR loops in other contexts fixed
+- Don't mark Robot Python variables as a failure
+- Fixed inspection of nested variables
 
 ## 0.5.5